Fix compilation for when uielements is part of the project
