+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/OnScreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: deaa7160c96d24ceabbdb2985165be8c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
