+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/OnScreen/OnScreenButtonEditor.cs
@@ -1,0 +1,36 @@
+#if PACKAGE_DOCS_GENERATION || UNITY_INPUT_SYSTEM_ENABLE_UI
+using UnityEngine.EventSystems;
+using UnityEngine.InputSystem.Layouts;
+using UnityEngine.InputSystem.OnScreen;
+
+using UnityEngine.InputSystem.Editor;
+using UnityEngine.InputSystem;
+
+namespace UnityEditor.InputSystem.OnScreen {
+    [UnityEditor.CustomEditor(typeof(OnScreenButton))]
+    internal class OnScreenButtonEditor : UnityEditor.Editor
+    {
+        private UnityEditor.SerializedProperty m_ControlPathInternal;
+
+        public void OnEnable()
+        {
+            m_ControlPathInternal = serializedObject.FindProperty(OnScreenButton.OnScreenButtonEditorHelper.m_ControlPath);
+        }
+
+        public void OnDisable()
+        {
+            new InputComponentEditorAnalytic(InputSystemComponent.OnScreenButton).Send();
+        }
+
+        public override void OnInspectorGUI()
+        {
+            // Current implementation has UGUI dependencies (ISXB-915, ISXB-916)
+            UGUIOnScreenControlEditorUtils.ShowWarningIfNotPartOfCanvasHierarchy((OnScreenButton)target);
+
+            UnityEditor.EditorGUILayout.PropertyField(m_ControlPathInternal);
+
+            serializedObject.ApplyModifiedProperties();
+        }
+    }
+}
+#endif
