+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/OnScreen/OnScreenButtonEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ec10af04a3b1d402f9693292ed3c663f
