+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/OnScreen/OnScreenStickEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b4fc1685f5c944313bcc95eb2a94bbdc
