+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c6210498d865a43b1b8e0a377f17ccee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
