+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/UI/VirtualMouseInputEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0537f205392654ccbaca8b1f35970e2b
