+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/InputSystemUIInputModule.cs
@@ -7,9 +7,6 @@ using UnityEngine.InputSystem.LowLevel;
 using UnityEngine.InputSystem.Utilities;
 using UnityEngine.Serialization;
 using UnityEngine.UI;
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
 
 ////FIXME: The UI is currently not reacting to pointers until they are moved after the UI module has been enabled. What needs to
 ////       happen is that point, trackedDevicePosition, and trackedDeviceOrientation have initial state checks. However, for touch,
@@ -1609,18 +1606,13 @@ namespace UnityEngine.InputSystem.UI
         }
 
 #if UNITY_EDITOR
+        internal static Action<InputSystemUIInputModule> m_Reset;
         /// <inheritdoc/>
         protected override void Reset()
         {
             base.Reset();
-
-            var asset = (InputActionAsset)AssetDatabase.LoadAssetAtPath(
-                UnityEngine.InputSystem.Editor.PlayerInputEditor.kDefaultInputActionsAssetPath,
-                typeof(InputActionAsset));
-            // Setting default asset and actions when creating via inspector
-            Editor.InputSystemUIInputModuleEditor.ReassignActions(this, asset);
+            m_Reset.Invoke(this);
         }
-
 #endif
 
         /// <inheritdoc/>
